Booking/backend/requirements.txt
fastapi==0.117.1
uvicorn==0.37.0
pydantic==2.11.9
python-multipart==0.0.20
python-dotenv==1.0.0
pymongo==3.13.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
